requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
pydub==0.25.1
numpy==1.26.4
pytest==8.0.2
python-multipart==0.0.9

# FIXED Whisper (Python 3.12 compatible)
openai-whisper @ git+https://github.com/openai/whisper.git

# Torch (must be installed properly)
torch==2.2.1

gTTS==2.5.1
httpx==0.27.2

# Fix for pkg_resources error
setuptools>=68.0.0
wheel